+++ pkg/engine/engine.go
@@ -31,7 +31,7 @@ type Rotator interface {
 	Archive(ctx context.Context, absBasePath, relArchivePathKey, archiveEntryPrefix string, toArchive metafile.MetafileInfo, p *pathrotation.Plan, timestampUTC time.Time) (metafile.MetafileInfo, error)
 	Stage(ctx context.Context, absBasePath, relStagePathKey, stageEntryPrefix string, toStage metafile.MetafileInfo, p *pathrotation.Plan, timestampUTC time.Time) (metafile.MetafileInfo, error)
 	Unstage(ctx context.Context, absBasePath string, stagedInfo metafile.MetafileInfo, p *pathrotation.Plan, timestampUTC time.Time) error
-	CleanupStagingPath(ctx context.Context, absBasePath, relStagePathKey string, p *pathrotation.Plan, timestampUTC time.Time) error
+	CleanupStage(ctx context.Context, absBasePath, relStagePathKey string, p *pathrotation.Plan, timestampUTC time.Time) error
 }
 
 // Retainer is an interface for the retention leaf package.